Add economy support. Update to 2.1.1
Not tested at all!

+++ src/com/drtshock/playervaults/commands/VaultOperations.java
@@ -7,6 +7,7 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
 import com.drtshock.playervaults.Main;
+import com.drtshock.playervaults.util.EconomyOperations;
 import com.drtshock.playervaults.util.Lang;
 import com.drtshock.playervaults.util.VaultManager;
 
@@ -23,6 +24,7 @@ public class VaultOperations {
 		}
 		return false;
 	}
+	
 	/*
 	 * TODO: Change how permissions are checked here.
 	 */
@@ -37,9 +39,14 @@ public class VaultOperations {
 				return false;
 			}
 			if(checkPerms(sender, number)) {
-				vm.loadVault(sender, sender.getName(), number);
-				sender.sendMessage(Lang.TITLE.toString() + Lang.OPEN_VAULT.toString().replace("%v", arg));
-				return true;
+				if(EconomyOperations.payToOpen(sender)) {
+					vm.loadVault(sender, sender.getName(), number);
+					sender.sendMessage(Lang.TITLE.toString() + Lang.OPEN_VAULT.toString().replace("%v", arg));
+					return true;
+				} else {
+					sender.sendMessage(Lang.TITLE.toString() + Lang.INSUFFICIENT_FUNDS);
+					return false;
+				}
 			} else {
 				Feedback.noPerms(sender);
 			}
@@ -48,7 +55,6 @@ public class VaultOperations {
 		}
 		return false;
 	}
-
 
 	public static boolean openOtherVault(Player sender, String user, String arg) {
 		if(sender.hasPermission("playervaults.admin")) {
@@ -72,6 +78,7 @@ public class VaultOperations {
 		}
 		return false;
 	}
+	
 	public static void deleteOwnVault(Player sender, String arg) {
 		if(arg.matches("^[0-9]{1,2}$")) {
 			int number = 0;
@@ -82,7 +89,10 @@ public class VaultOperations {
 				sender.sendMessage(Lang.TITLE.toString() + ChatColor.RED + Lang.MUST_BE_NUMBER);
 			}
 			try {
-				vm.deleteVault(sender, sender.getName(), number);
+				if(EconomyOperations.refundOnDelete(sender)) {
+					vm.deleteVault(sender, sender.getName(), number);
+					return;
+				}
 			} catch (IOException e) {
 				sender.sendMessage(Lang.TITLE.toString() + Lang.DELETE_VAULT_ERROR);
 			}
@@ -90,6 +100,7 @@ public class VaultOperations {
 			sender.sendMessage(Lang.TITLE.toString()+ Lang.MUST_BE_NUMBER);
 		}
 	}
+	
 	public static void deleteOtherVault(CommandSender sender, String user, String arg) {
 		if(sender.hasPermission("playervaults.delete")) {
 			if(arg.matches("^[0-9]{1,2}$")) {
